+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # valantic javascript utils
 
 ## unreleased
+
+- Moved all helper functions from the vue-template repo to this one and added tests for them.